refactor: reorganize and encapsulate ROS2 message handling functionality
Streamline the handling of ROS2 messages by encapsulating related functionality into a new utility header file. This enhances code clarity and maintainability while adhering to best practices in modular design.
